fix: flush read buffer before writing request
If a device is slow responding (for instance due to a slow network
connection) you might go to send the next poll and immediately receive
the previous response.  The result is a mismatch in the transaction ID
from which the connection will never recover.  There are a few ways to
mitigate this, including receiving the response in a loop, but since
this modbus tcp implementation is completely synchronous the simplest
thing to do is to simply flush the receive buffer immediately prior to
sending the next poll.