feat(cmd/modbus-cli): init cli tool for modbus queries
This pr provides a modbus-cli tool that can be used to check/modify modbus
registers manually or debug modbus devices.
Currently supported are the modbus functions 0x03, 0x04, 0x06 and 0x10 for TCP
and RTU. When reading and writing, a distinction can be made between the byte
orders BigEndian and LittleEndian, independent of the operation.