fix(tcp): falsy comparison against nil net.Conn
When the result of net.Dial() is assigned to 'mb.conn' unconditionally,
'mb.conn' will have the value 'net.Conn(nil)' instead of an explicit nil
value[1] when the function fails (returns an error). As a result,
comparisons to nil are falsy and panics may occur due to nil pointer
dereferences.

[1]: https://100go.co/#returning-a-nil-receiver-45